chore: ensure input blur doesn't close input, but instead overlay click or esc key
