+++ src/components/dropdowns/attachment-dropdowns.tsx
@@ -4,10 +4,10 @@ import { ScreenshotDropdown } from './screenshot/screenshot-dropdown'
 
 export function AttachmentDropdowns() {
   return (
-    <div className="flex items-center gap-2">
+    <>
       <ClipboardFileDropdown />
       <ScreenshotDropdown />
       <ConversationsDropdown />
-    </div>
+    </>
   )
 }
